Split query serialization
This creates a new file, query_serializer.c, which contains actual query
serialization routines.
